+++ app/src/main/java/com/example/drivingefficiencyapp/ui/MainMenuActivity.kt
@@ -14,6 +14,7 @@ import android.content.Intent
 import android.os.Bundle
 import androidx.appcompat.app.AppCompatActivity
 import com.example.drivingefficiencyapp.databinding.MainMenuActivityBinding
+import com.example.drivingefficiencyapp.profile.ProfileActivity
 import com.example.drivingefficiencyapp.trip.TripsActivity
 
 class MainMenuActivity : AppCompatActivity() {